requirements.txt
alabaster==0.7.12
amqp==2.5.0
astroid==2.2.5
Babel==2.7.0
billiard==3.6.0.0
boto3==1.9.171
botocore==1.12.171
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Collectfast==0.6.2
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.2
django-anymail==6.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.1.0
django-log-request-id==1.3.2
django-rest-swagger==2.2.0
django-secure==1.0.1
django-storages==1.7.1
djangorestframework==3.9.4
dnspython==1.16.0
docutils==0.14
eventlet==0.25.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
isort==4.3.20
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.3
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
mysqlclient==1.4.2.post1
openapi-codec==1.3.2
packaging==19.0
Pygments==2.4.2
pylint==2.3.1
pylint-celery==0.3
pylint-django==2.0.8
pylint-plugin-utils==0.5
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.1.2
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wrapt==1.11.2
gevent==1.4.0